Second Commit

+++ src/com/mycompany/basicmathoperations/simple/Multiply.java
@@ -12,7 +12,7 @@ public class Multiply {
 	 * To multiply 2 int numbers
 	 * @return num1 * num2
 	 */
-	public int multiply(int num1, int num2) {
+	public float multiply(float num1, float num2) {
 		return num1 * num2;
 	}
 }
